Add passing tests to API and Fetch
