+++ src/utilities/test/API.test.js
@@ -23,31 +23,6 @@ describe('API', () => {
    });
  });
 
- describe('fetchCorrectCountry', () => {
-     beforeEach(() => {
-     })
-     it.skip('should call fetchCountries with the correct URL', async () => {})
-     it.skip('should return a resolved object', async () => {})
- })
-
- describe('fetchAllCountries', () => {
- 		let url;
-     beforeEach(() => {
-				url = 'https://flagz4u.herokuapp.com/api/v1/country';
-     })
-
-     it.skip('should call fetchCountries with the correct URL', async () => {
-     		API.fetchAllCountries()
-     		expect(API.fetchData).toHaveBeenCalledWith(url)
-     })
-
-     it.skip('should return a resolved array', async () => {
-     		const countries = await API.fetchAllCountries();
-     		expect(countries).toEqual([])
-     })
- })
-
-
  describe('loginUser', () => {
  		let mockUser;
      beforeEach(() => {
